+++ .gitignore
@@ -2,15 +2,5 @@
 node_modules
 
 # config
-/config/nessie.json
-/config/database.js
-/config/ca-certificate.cer
-
-#database
-/database/nessie.db
-/database/nessie.db-journal
-/database/nessie_backup.db
-
-# system files
-.DS_Store
 .env
+src/config
