+++ src/services/adapters.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-const { apiKey } = require('../../config/nessie.json');
+const { apiKey } = require('../config/nessie.json');
 
 //Documentation on API: https://apexlegendsapi.com/documentation.php
 const url = `https://api.mozambiquehe.re/maprotation?version=2&auth=${apiKey}`;
