+++ src/services/database.js
@@ -1,7 +1,7 @@
 const { Pool } = require('pg');
-const { databaseConfig } = require('../../config/database');
+const { databaseConfig } = require('../config/database');
 const { sendGuildUpdateNotification } = require('../utils/helpers');
-const { defaultPrefix } = require('../../config/nessie.json');
+const { defaultPrefix } = require('../config/nessie.json');
 
 exports.pool = new Pool(databaseConfig); //Intialise pool to connect to our cloud database; more information https://node-postgres.com/features/pooling
 /**
